move (back) crawl and friends from coerce.js -> plot_schema.js
- (now that this doesn't cause circular deps anymore)
